ai-career-agent/requirements.txt
Flask==3.1.0
python-dotenv==1.0.1
pytest==8.3.4
google-genai